+++ src/component/HeroSection.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Box } from '@chakra-ui/react';
-import CTA from '../HOC/CTA';
+import CTA from '../HOC/TextDescription';
 import Logo from '../svg_components/Logo';
 
 const data = {
